+++ packages/common/src/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @copyright   2017-2019, Miles Johnson
+ * @license     https://opensource.org/licenses/MIT
+ */
+
+import { Blueprint, Predicates } from 'optimal';
+import Optionable from './Optionable';
+
+export { Optionable, Blueprint, Predicates };