+++ packages/core/src/Plugin.ts
@@ -1,33 +1,17 @@
-import optimal, { predicates, Blueprint, Predicates } from 'optimal';
+import { Optionable } from '@boost/common';
 import Tool from './Tool';
 
 export const DEFAULT_PLUGIN_PRIORITY = 100;
 
-export default class Plugin<Options extends object = {}> {
+export default abstract class Plugin<Options extends object = {}> extends Optionable<Options> {
   moduleName: string = '';
 
   name: string = '';
-
-  options: Required<Options>;
 
   priority: number = DEFAULT_PLUGIN_PRIORITY;
 
   // @ts-ignore Set after instantiation
   tool: Tool<any>;
-
-  constructor(options?: Options) {
-    this.options = optimal({ ...options }, this.blueprint(predicates), {
-      name: this.constructor.name,
-    });
-  }
-
-  /**
-   * Define an optimal blueprint in which to validate and build the
-   * options passed to the constructor.
-   */
-  blueprint(preds: Predicates): Blueprint<Required<Options>> {
-    return {} as any;
-  }
 
   /**
    * Called once the plugin has been loaded by the tool.