+++ scripts/build-packages.sh
@@ -12,6 +12,7 @@ build_pkg() {
   node ../../node_modules/.bin/tsc
 }
 
+build_pkg "./packages/common"
 build_pkg "./packages/event"
 build_pkg "./packages/core"
 build_pkg "./packages/reporter-nyan"