+++ tests/__fixtures__/plugin-exported-instance/index.js
@@ -2,4 +2,10 @@ const path = require('path');
 
 const Plugin = require(path.join(__dirname, '../../../packages/core/src/Plugin')).default;
 
-module.exports = new Plugin();
+class TestPlugin extends Plugin {
+  blueprint() {
+    return {};
+  }
+}
+
+module.exports = new TestPlugin();